Account, Homepage changes
Made it so you go to "timeline" instead of "my account" when you log in. I gave the "my account" its own navbar on the left. I also made it so you don't need to type your username to make a post or comment, but it picks the username of the account you're logged in to.